Refactor. Login.jsx made it modulair, made 1 extra component with formInputs and made login a helper (useLogin). This way the code keeps cleaner.
