+++ src/helper/useLogin.js
@@ -1,0 +1,34 @@
+import { useState } from 'react';
+
+export const useLogin = () => {
+    const [formData, setFormData] = useState({ username: '', password: '' });
+
+    const handleInputChange = (event) => {
+        const { name, value } = event.target;
+        setFormData({ ...formData, [name]: value });
+    };
+
+    const handleSubmit = async (event) => {
+        event.preventDefault();
+        try {
+            const response = await fetch('http://localhost:8080/authenticate', {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(formData),
+            });
+            if (response.ok) {
+                const { jwt } = await response.json();
+                localStorage.setItem('jwtToken', jwt);
+                window.location.href = '/profile';
+            } else {
+                const errorData = await response.json();
+                alert(`Login failed: ${errorData.message}`);
+            }
+        } catch (error) {
+            console.error('Error logging in:', error);
+            alert('Login failed. Please try again.');
+        }
+    };
+
+    return { formData, handleInputChange, handleSubmit };
+};
